requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic[email]
python-dotenv
pytest
pytest-asyncio
httpx
aiosqlite
greenlet